Refactorized the method getValue() for connection items in  Builder project
