+++ IntroOOP/Builder/src/com/globant/builder/connection/ConnectionItem.java
@@ -1,11 +1,20 @@
 package com.globant.builder.connection;
 
+import java.util.Scanner;
+
 public abstract class ConnectionItem {
 
   // Connection item field
   protected String value="";
 
   // Retrieves info
-  public abstract String getValue();
+  public String getValue() {
+    if (this.value.equals("")) {
+      Scanner input = new Scanner(System.in);
 
+      System.out.println("Enter a valid"+this.getClass().getName());
+      this.value = input.nextLine();
+    }
+    return this.value;
+  }
 }
