+++ IntroOOP/Builder/src/com/globant/builder/connection/URL.java
@@ -4,15 +4,7 @@ import java.util.Scanner;
 
 public class URL extends ConnectionItem {
 
-  @Override
-  public String getValue() {
-    if (this.value.equals("")) {
-      Scanner input = new Scanner(System.in);
-      System.out.println("Enter a valid URL:");
-      this.value = input.nextLine();
-    }
-    return this.value;
-  }
+
 
 
 }
